app/requirements.txt
aiofiles==23.1.0
alembic==1.11.1
amqp==5.1.1
anyio==3.7.0
async-timeout==4.0.2
Babel==2.12.1
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
fastapi==0.97.0
fastapi-babel==0.0.8
FastAPI-SQLAlchemy==0.2.1
fastapi-users==11.0.0
flower==1.2.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
kombu==5.3.0
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
minio==7.1.15
numpy==1.24.3
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
pyotp==2.8.0
pytest==7.3.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
redis==4.5.5
rsa==4.9
sentry-sdk==1.25.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
requests
